+++ cg/apps/tb/api.py
@@ -5,6 +5,7 @@ import logging
 from typing import Any, Optional
 
 import requests
+from cg.constants import Pipeline
 from google.auth import jwt
 from google.auth.crypt import RSASigner
 
@@ -17,8 +18,8 @@ LOG = logging.getLogger(__name__)
 class TrailblazerAPI:
     """Interface to Trailblazer for `cg`."""
 
-    __STARTED_STATUSES = ["completed", "failed", "pending", "running"]
-    __ONGOING_STATUSES = ["pending", "running"]
+    __STARTED_STATUSES = ["completed", "failed", "pending", "running", "error"]
+    __ONGOING_STATUSES = ["pending", "running", "error"]
 
     def __init__(self, config: dict):
         self.service_account = config["trailblazer"]["service_account"]
@@ -60,7 +61,7 @@ class TrailblazerAPI:
         before: dt.datetime = None,
         is_visible: bool = None,
         family: str = None,
-        data_analysis: str = None,
+        data_analysis: Pipeline = None,
     ) -> list:
         request_body = {
             "case_id": case_id,
@@ -71,7 +72,7 @@ class TrailblazerAPI:
             "before": str(before) if before else None,
             "is_visible": is_visible,
             "family": family,
-            "data_analysis": data_analysis,
+            "data_analysis": str(data_analysis).upper() if data_analysis else None,
         }
         response = self.query_trailblazer(command="query-analyses", request_body=request_body)
         if response:
@@ -135,7 +136,7 @@ class TrailblazerAPI:
         out_dir: str,
         priority: str,
         email: str = None,
-        data_analysis: str = None,
+        data_analysis: Pipeline = None,
     ) -> TrailblazerAnalysis:
         request_body = {
             "case_id": case_id,
@@ -144,7 +145,7 @@ class TrailblazerAPI:
             "config_path": config_path,
             "out_dir": out_dir,
             "priority": priority,
-            "data_analysis": data_analysis,
+            "data_analysis": str(data_analysis).upper(),
         }
         response = self.query_trailblazer(command="add-pending-analysis", request_body=request_body)
         if response: