+++ cg/cli/upload/delivery_report.py
@@ -8,6 +8,7 @@ import click
 
 from cg.apps.hk import HousekeeperAPI
 from cg.apps.scoutapi import ScoutAPI
+from cg.constants import Pipeline
 from cg.exc import DeliveryReportError, CgError
 
 from .utils import suggest_cases_delivery_report
@@ -33,7 +34,9 @@ def delivery_reports(context, print_console, force_report):
     click.echo(click.style("----------------- DELIVERY REPORTS ------------------------"))
 
     exit_code = SUCCESS
-    for analysis_obj in context.obj["status_db"].analyses_to_delivery_report(pipeline="mip"):
+    for analysis_obj in context.obj["status_db"].analyses_to_delivery_report(
+        pipeline=Pipeline.MIP_DNA
+    ):
         case_id = analysis_obj.family.internal_id
         LOG.info("Uploading delivery report for case: %s", case_id)
         try: