+++ cg/meta/orders/schema.py
@@ -1,6 +1,5 @@
 from collections import Iterable
-from enum import Enum
-
+from cg.utils.StrEnum import StrEnum
 from pyschemes import Scheme, validators
 
 from cg.constants import (
@@ -9,18 +8,19 @@ from cg.constants import (
     STATUS_OPTIONS,
     CAPTUREKIT_OPTIONS,
     CONTAINER_OPTIONS,
+    Pipeline,
 )
 
 
-class OrderType(Enum):
-    MIP_RNA = "mip_rna"
-    EXTERNAL = "external"
-    FASTQ = "fastq"
-    RML = "rml"
-    MIP = "mip"
-    MICROBIAL = "microbial"
-    METAGENOME = "metagenome"
-    BALSAMIC = "balsamic"
+class OrderType(StrEnum):
+    BALSAMIC: str = str(Pipeline.BALSAMIC)
+    EXTERNAL: str = "external"
+    FASTQ: str = "fastq"
+    METAGENOME: str = "metagenome"
+    MICROSALT: str = str(Pipeline.MICROSALT)
+    MIP_DNA: str = str(Pipeline.MIP_DNA)
+    MIP_RNA: str = str(Pipeline.MIP_RNA)
+    RML: str = "rml"
 
 
 class ListValidator(validators.Validator):
@@ -95,6 +95,7 @@ MIP_SAMPLE = {
     "name": validators.RegexValidator(NAME_PATTERN),
     # customer
     "data_analysis": str,
+    "data_delivery": OptionalNone(TypeValidatorNone(str)),
     "application": str,
     "family_name": validators.RegexValidator(NAME_PATTERN),
     "sex": OptionalNone(validators.Any(SEX_OPTIONS)),
@@ -134,6 +135,7 @@ BALSAMIC_SAMPLE = {
     "name": validators.RegexValidator(NAME_PATTERN),
     "container": OptionalNone(validators.Any(CONTAINER_OPTIONS)),
     "data_analysis": str,
+    "data_delivery": OptionalNone(TypeValidatorNone(str)),
     "application": str,
     "sex": OptionalNone(validators.Any(SEX_OPTIONS)),
     "family_name": validators.RegexValidator(NAME_PATTERN),
@@ -164,6 +166,7 @@ MIP_RNA_SAMPLE = {
     "name": validators.RegexValidator(NAME_PATTERN),
     # customer
     "data_analysis": str,
+    "data_delivery": OptionalNone(TypeValidatorNone(str)),
     "application": str,
     "family_name": validators.RegexValidator(NAME_PATTERN),
     "sex": OptionalNone(validators.Any(SEX_OPTIONS)),
@@ -181,7 +184,7 @@ MIP_RNA_SAMPLE = {
     "comment": OptionalNone(TypeValidatorNone(str)),
     # Orderform 1508:19
     "from_sample": validators.RegexValidator(NAME_PATTERN),
-    "time_point": TypeValidatorNone(str),
+    "time_point": OptionalNone(TypeValidatorNone(str)),
 }
 
 EXTERNAL_SAMPLE = {
@@ -189,6 +192,7 @@ EXTERNAL_SAMPLE = {
     # Order portal specific
     "internal_id": OptionalNone(TypeValidatorNone(str)),
     "data_analysis": str,
+    "data_delivery": OptionalNone(TypeValidatorNone(str)),
     # "required for new samples"
     "name": validators.RegexValidator(NAME_PATTERN),
     "capture_kit": OptionalNone(TypeValidatorNone(str)),
@@ -218,6 +222,7 @@ FASTQ_SAMPLE = {
     "name": validators.RegexValidator(NAME_PATTERN),
     "container": OptionalNone(validators.Any(CONTAINER_OPTIONS)),
     "data_analysis": str,
+    "data_delivery": OptionalNone(TypeValidatorNone(str)),
     "application": str,
     "sex": OptionalNone(validators.Any(SEX_OPTIONS)),
     # todo: implement in OP or remove from OF
@@ -250,6 +255,7 @@ RML_SAMPLE = {
     "pool": str,
     "application": str,
     "data_analysis": str,
+    "data_delivery": OptionalNone(TypeValidatorNone(str)),
     "volume": str,
     "concentration": str,
     "concentration_sample": str,
@@ -265,13 +271,14 @@ RML_SAMPLE = {
     "capture_kit": OptionalNone(validators.Any(CAPTUREKIT_OPTIONS)),
 }
 
-MICROBIAL_SAMPLE = {
+MICROSALT_SAMPLE = {
     # 1603:6 Orderform Microbial WGS
     # "These fields are required"
     "name": validators.RegexValidator(NAME_PATTERN),
     "organism": str,
     "reference_genome": str,
     "data_analysis": str,
+    "data_delivery": OptionalNone(TypeValidatorNone(str)),
     "application": str,
     "require_qcok": bool,
     "elution_buffer": str,
@@ -295,6 +302,7 @@ METAGENOME_SAMPLE = {
     "name": validators.RegexValidator(NAME_PATTERN),
     "container": OptionalNone(validators.Any(CONTAINER_OPTIONS)),
     "data_analysis": str,
+    "data_delivery": OptionalNone(TypeValidatorNone(str)),
     "application": str,
     "require_qcok": bool,
     "elution_buffer": str,
@@ -314,7 +322,7 @@ ORDER_SCHEMES = {
     OrderType.EXTERNAL: Scheme(
         {**BASE_PROJECT, "samples": ListValidator(EXTERNAL_SAMPLE, min_items=1)}
     ),
-    OrderType.MIP: Scheme({**BASE_PROJECT, "samples": ListValidator(MIP_SAMPLE, min_items=1)}),
+    OrderType.MIP_DNA: Scheme({**BASE_PROJECT, "samples": ListValidator(MIP_SAMPLE, min_items=1)}),
     OrderType.BALSAMIC: Scheme(
         {**BASE_PROJECT, "samples": ListValidator(BALSAMIC_SAMPLE, min_items=1)}
     ),
@@ -323,8 +331,8 @@ ORDER_SCHEMES = {
     ),
     OrderType.FASTQ: Scheme({**BASE_PROJECT, "samples": ListValidator(FASTQ_SAMPLE, min_items=1)}),
     OrderType.RML: Scheme({**BASE_PROJECT, "samples": ListValidator(RML_SAMPLE, min_items=1)}),
-    OrderType.MICROBIAL: Scheme(
-        {**BASE_PROJECT, "samples": ListValidator(MICROBIAL_SAMPLE, min_items=1)}
+    OrderType.MICROSALT: Scheme(
+        {**BASE_PROJECT, "samples": ListValidator(MICROSALT_SAMPLE, min_items=1)}
     ),
     OrderType.METAGENOME: Scheme(
         {**BASE_PROJECT, "samples": ListValidator(METAGENOME_SAMPLE, min_items=1)}