+++ cg/meta/orders/status.py
@@ -2,6 +2,7 @@
 import datetime as dt
 from typing import List
 
+from cg.constants import Pipeline, DataDelivery
 from cg.exc import OrderError
 from cg.store import models
 
@@ -33,21 +34,22 @@ class StatusHandler:
         pools = {}
 
         for sample in data["samples"]:
-            name = sample["pool"]
+            pool_name = sample["pool"]
             application = sample["application"]
             data_analysis = sample["data_analysis"]
+            data_delivery = sample.get("data_delivery")
             capture_kit = sample.get("capture_kit")
 
-            if name not in pools:
-                pools[name] = {}
-                pools[name]["name"] = name
-                pools[name]["applications"] = set()
-                pools[name]["capture_kits"] = set()
+            if pool_name not in pools:
+                pools[pool_name] = {}
+                pools[pool_name]["name"] = pool_name
+                pools[pool_name]["applications"] = set()
+                pools[pool_name]["capture_kits"] = set()
 
-            pools[name]["applications"].add(application)
+            pools[pool_name]["applications"].add(application)
 
             if capture_kit:
-                pools[name]["capture_kits"].add(capture_kit)
+                pools[pool_name]["capture_kits"].add(capture_kit)
 
         # each pool must only have one application type
         for pool in pools.values():
@@ -82,6 +84,7 @@ class StatusHandler:
                     "name": pool_name,
                     "application": application,
                     "data_analysis": data_analysis,
+                    "data_delivery": data_delivery,
                     "capture_kit": capture_kit,
                 }
             )
@@ -98,6 +101,7 @@ class StatusHandler:
                     "application": sample["application"],
                     "comment": sample.get("comment"),
                     "data_analysis": sample["data_analysis"],
+                    "data_delivery": sample.get("data_delivery"),
                     "internal_id": sample.get("internal_id"),
                     "name": sample["name"],
                     "priority": sample["priority"],
@@ -123,6 +127,7 @@ class StatusHandler:
                 {
                     "application": sample_data["application"],
                     "comment": sample_data.get("comment"),
+                    "data_delivery": sample_data.get("data_delivery"),
                     "internal_id": sample_data.get("internal_id"),
                     "name": sample_data["name"],
                     "organism_id": sample_data["organism"],
@@ -141,14 +146,13 @@ class StatusHandler:
         cases = cls.group_cases(data["samples"])
 
         for case_name, case_samples in cases.items():
-            values = set(sample.get("priority", "standard") for sample in case_samples)
-            if len(values) > 1:
-                raise ValueError(f"different 'priority' values: {case_name} - {values}")
-            priority = values.pop()
+            priority = cls.get_single_value(case_name, case_samples, "priority", "standard")
+            data_analysis = cls.get_single_value(case_name, case_samples, "data_analysis")
+
             panels = set(panel for sample in case_samples for panel in sample.get("panels", set()))
             case = {
                 # Set from first sample until order portal sets this on case level
-                "data_analysis": case_samples[0]["data_analysis"],
+                "data_analysis": data_analysis,
                 "name": case_name,
                 "priority": priority,
                 "panels": list(panels),
@@ -157,6 +161,7 @@ class StatusHandler:
                         "application": sample["application"],
                         "capture_kit": sample.get("capture_kit"),
                         "comment": sample.get("comment"),
+                        "data_delivery": sample.get("data_delivery"),
                         "father": sample.get("father"),
                         "from_sample": sample.get("from_sample"),
                         "internal_id": sample.get("internal_id"),
@@ -165,7 +170,7 @@ class StatusHandler:
                         "sex": sample["sex"],
                         "status": sample.get("status"),
                         "time_point": sample.get("time_point"),
-                        "tumour": sample.get("tumour") or False,
+                        "tumour": sample.get("tumour", False),
                     }
                     for sample in case_samples
                 ],
@@ -173,6 +178,14 @@ class StatusHandler:
 
             status_data["families"].append(case)
         return status_data
+
+    @classmethod
+    def get_single_value(cls, case_name, case_samples, value_key, value_default=None):
+        values = set(sample.get(value_key, value_default) for sample in case_samples)
+        if len(values) > 1:
+            raise ValueError(f"different sample {value_key} values: {case_name} - {values}")
+        single_value = values.pop()
+        return single_value
 
     def store_cases(
         self, customer: str, order: str, ordered: dt.datetime, ticket: int, cases: List[dict]
@@ -188,7 +201,7 @@ class StatusHandler:
                 case_obj.panels = case["panels"]
             else:
                 case_obj = self.status.add_family(
-                    data_analysis=case["data_analysis"],
+                    data_analysis=Pipeline(case["data_analysis"]),
                     name=case["name"],
                     panels=case["panels"],
                     priority=case["priority"],
@@ -205,6 +218,7 @@ class StatusHandler:
                     new_sample = self.status.add_sample(
                         capture_kit=sample["capture_kit"],
                         comment=sample["comment"],
+                        data_delivery=sample["data_delivery"],
                         from_sample=sample["from_sample"],
                         internal_id=sample["internal_id"],
                         name=sample["name"],
@@ -263,14 +277,15 @@ class StatusHandler:
         with self.status.session.no_autoflush:
             for sample in samples:
                 new_sample = self.status.add_sample(
-                    name=sample["name"],
+                    comment=sample["comment"],
+                    data_delivery=sample["data_delivery"],
                     internal_id=sample["internal_id"],
-                    sex=sample["sex"] or "unknown",
+                    name=sample["name"],
                     order=order,
                     ordered=ordered,
-                    ticket=ticket,
                     priority=sample["priority"],
-                    comment=sample["comment"],
+                    sex=sample["sex"] or "unknown",
+                    ticket=ticket,
                     tumour=sample["tumour"],
                 )
                 new_sample.customer = customer_obj
@@ -282,7 +297,7 @@ class StatusHandler:
                 new_samples.append(new_sample)
 
                 new_family = self.status.add_family(
-                    data_analysis=sample["data_analysis"],
+                    data_analysis=Pipeline(sample["data_analysis"]),
                     name=sample["name"],
                     panels=None,
                     priority=sample["priority"],
@@ -311,6 +326,7 @@ class StatusHandler:
         with self.status.session.no_autoflush:
             for sample in samples:
                 new_sample = self.status.add_sample(
+                    data_delivery=str(DataDelivery.FASTQ),
                     name=sample["name"],
                     internal_id=sample["internal_id"],
                     sex=sample["sex"] or "unknown",
@@ -322,28 +338,24 @@ class StatusHandler:
                     tumour=sample["tumour"],
                 )
                 new_sample.customer = customer_obj
-
                 application_tag = sample["application"]
                 application_version = self.status.current_application_version(application_tag)
                 if application_version is None:
                     raise OrderError(f"Invalid application: {sample['application']}")
                 new_sample.application_version = application_version
                 new_samples.append(new_sample)
-
                 new_family = self.status.add_family(
-                    data_analysis="fastq",
+                    data_analysis=Pipeline.MIP_DNA,
                     name=sample["name"],
                     panels=["OMIM-AUTO"],
                     priority="research",
                 )
                 new_family.customer = production_customer
                 self.status.add(new_family)
-
                 new_relationship = self.status.relate_sample(
                     family=new_family, sample=new_sample, status=sample["status"] or "unknown"
                 )
                 self.status.add(new_relationship)
-
                 new_delivery = self.status.add_delivery(destination="caesar", sample=new_sample)
                 self.status.add(new_delivery)
 
@@ -354,7 +366,7 @@ class StatusHandler:
         self,
         comment: str,
         customer: str,
-        data_analysis: str,
+        data_analysis: Pipeline,
         order: str,
         ordered: dt.datetime,
         samples: List[dict],
@@ -406,6 +418,7 @@ class StatusHandler:
                     application_version=application_version,
                     comment=sample_data["comment"],
                     customer=customer_obj,
+                    data_delivery=sample_data["data_delivery"],
                     internal_id=sample_data["internal_id"],
                     name=sample_data["name"],
                     order=order,
@@ -448,7 +461,7 @@ class StatusHandler:
                 ordered=ordered,
                 ticket=ticket,
                 application_version=application_version,
-                data_analysis=pool["data_analysis"],
+                data_analysis=Pipeline(pool["data_analysis"]),
                 capture_kit=pool["capture_kit"],
             )
             new_delivery = self.status.add_delivery(destination="caesar", pool=new_pool)