+++ cg/meta/upload/scoutapi.py
@@ -2,6 +2,7 @@
 
 import logging
 from pathlib import Path
+from datetime import datetime
 
 import requests
 from ruamel import yaml
@@ -12,6 +13,9 @@ from cg.apps.lims import LimsAPI
 from cg.apps.madeline.api import MadelineAPI
 from cg.meta.workflow.mip import MipAnalysisAPI
 from cg.store import models
+from cg.models.scout_load_config import ScoutCase, ScoutIndividual
+
+from housekeeper.store import models as hk_models
 
 LOG = logging.getLogger(__name__)
 
@@ -44,8 +48,9 @@ class UploadScoutAPI:
 
     def build_samples(self, analysis_obj: models.Analysis, hk_version_id: int = None):
         """Loop over the samples in an analysis and build dicts from them"""
-
+        LOG.info("Building samples")
         for link_obj in analysis_obj.family.links:
+            LOG.debug("Found link obj %ss", link_obj)
             sample_id = link_obj.sample.internal_id
             bam_path = self.fetch_file_path("bam", sample_id, hk_version_id)
             alignment_file_path = self.fetch_file_path("cram", sample_id, hk_version_id)
@@ -74,22 +79,33 @@ class UploadScoutAPI:
                 "tissue_type": lims_sample.get("source", "unknown"),
                 "vcf2cytosure": vcf2cytosure_path,
             }
-            yield sample
+            scout_sample = ScoutIndividual(**sample)
+            yield scout_sample.dict(exclude_none=True)
 
-    def generate_config(self, analysis_obj: models.Analysis) -> dict:
+    def generate_config(self, analysis_obj: models.Analysis, rank_score_threshold: int = 5) -> dict:
         """Fetch data about an analysis to load Scout."""
-        analysis_date = analysis_obj.started_at or analysis_obj.completed_at
-        hk_version = self.housekeeper.version(analysis_obj.family.internal_id, analysis_date)
-        analysis_data = self.analysis.get_latest_metadata(analysis_obj.family.internal_id)
+        LOG.info("Generate scout load config")
+        analysis_date: datetime = analysis_obj.started_at or analysis_obj.completed_at
+        LOG.debug("Found analysis date %s", analysis_date)
+        # Fetch last version from housekeeper
+        # This should be safe since analyses are only added if data is analysed
+        hk_version: hk_models.Version = self.housekeeper.last_version(
+            analysis_obj.family.internal_id
+        )
+        LOG.debug("Found housekeeper version %s", hk_version.id)
+        analysis_data: dict = self.analysis.get_latest_metadata(analysis_obj.family.internal_id)
+        genome_build = "38" if "38" in analysis_data.get("genome_build", "") else "37"
         data = {
             "analysis_date": analysis_obj.completed_at,
             "default_gene_panels": analysis_obj.family.panels,
             "family": analysis_obj.family.internal_id,
             "family_name": analysis_obj.family.name,
+            "rank_score_threshold": rank_score_threshold,
             "gene_panels": self.analysis.convert_panels(
                 analysis_obj.family.customer.internal_id, analysis_obj.family.panels
-            ),
-            "human_genome_build": analysis_data.get("genome_build"),
+            )
+            or None,
+            "human_genome_build": genome_build,
             "owner": analysis_obj.family.customer.internal_id,
             "rank_model_version": analysis_data.get("rank_model_version"),
             "samples": list(),
@@ -97,7 +113,6 @@ class UploadScoutAPI:
         }
         for sample in self.build_samples(analysis_obj, hk_version.id):
             data["samples"].append(sample)
-
         self._include_mandatory_files(data, hk_version)
         self._include_peddy_files(data, hk_version)
         self._include_optional_files(data, hk_version)
@@ -111,7 +126,11 @@ class UploadScoutAPI:
         else:
             LOG.info("family of 1 sample - skip pedigree graph")
 
-        return data
+        # Validate that the config is correct
+
+        scout_case = ScoutCase(**data)
+
+        return scout_case.dict(exclude_none=True)
 
     @staticmethod
     def get_load_config_tag() -> str: