+++ tests/cli/test_cli_add_family.py
@@ -1,9 +1,10 @@
 """This script tests the cli methods to add families to status-db"""
 from datetime import datetime
 
+from cg.constants import Pipeline
 from cg.store import Store
 
-ANALYSIS = "mip"
+CLI_OPTION_ANALYSIS = Pipeline.BALSAMIC
 
 
 def test_add_family_required(invoke_cli, disk_store: Store):
@@ -25,7 +26,7 @@ def test_add_family_required(invoke_cli, disk_store: Store):
             "--panel",
             panel_id,
             "--analysis",
-            ANALYSIS,
+            CLI_OPTION_ANALYSIS,
             customer_id,
             name,
         ]
@@ -56,7 +57,7 @@ def test_add_family_bad_customer(invoke_cli, disk_store: Store):
             "--panel",
             panel_id,
             "--analysis",
-            ANALYSIS,
+            CLI_OPTION_ANALYSIS,
             customer_id,
             name,
         ]
@@ -85,7 +86,7 @@ def test_add_family_bad_panel(invoke_cli, disk_store: Store):
             "--panel",
             panel_id,
             "--analysis",
-            ANALYSIS,
+            CLI_OPTION_ANALYSIS,
             customer_id,
             name,
         ]
@@ -119,7 +120,7 @@ def test_add_family_priority(invoke_cli, disk_store: Store):
             "--priority",
             priority,
             "--analysis",
-            ANALYSIS,
+            CLI_OPTION_ANALYSIS,
             customer_id,
             name,
         ]