+++ tests/cli/upload/test_cli_upload_auto.py
@@ -2,6 +2,7 @@
 import logging
 
 from cg.cli.upload.base import auto
+from cg.constants import Pipeline
 
 
 def test_upload_auto_with_pipeline_as_argument(
@@ -9,12 +10,12 @@ def test_upload_auto_with_pipeline_as_argument(
 ):
     """Test upload auto"""
     # GIVEN a store with a MIP analysis
-    helpers.add_analysis(store=sample_store, completed_at=timestamp, pipeline="MIP")
+    pipeline = Pipeline.MIP_DNA
+    helpers.add_analysis(store=sample_store, completed_at=timestamp, pipeline=pipeline)
 
     # WHEN uploading all analysis from pipeline MIP
     caplog.set_level(logging.INFO)
-    cli_runner.invoke(auto, ["--pipeline", "MIP"], obj=base_context)
+    cli_runner.invoke(auto, ["--pipeline", "mip-dna"], obj=base_context)
 
     # THEN assert that the MIP analysis was successfully uploaded
-    with caplog.at_level(logging.INFO):
-        assert "uploading family" in caplog.text
+    assert "Uploading family" in caplog.text