+++ tests/conftest.py
@@ -15,6 +15,7 @@ import ruamel.yaml
 from cg.apps.hk import HousekeeperAPI
 from cg.apps.mip import parse_sampleinfo, parse_qcmetrics
 from cg.apps.gt import GenotypeAPI
+from cg.constants import Pipeline
 from cg.models import CompressionData
 from cg.store import Store
 
@@ -70,7 +71,7 @@ def fixture_analysis_family_single(case_id: str, family_name: str, ticket_nr: in
     family = {
         "name": family_name,
         "internal_id": case_id,
-        "data_analysis": "mip",
+        "data_analysis": str(Pipeline.MIP_DNA),
         "application_type": "wgs",
         "panels": ["IEM", "EP"],
         "samples": [
@@ -94,7 +95,7 @@ def fixture_analysis_family(case_id: str, family_name: str, ticket_nr: int) -> d
     family = {
         "name": family_name,
         "internal_id": case_id,
-        "data_analysis": "mip",
+        "data_analysis": str(Pipeline.MIP_DNA),
         "application_type": "wgs",
         "panels": ["IEM", "EP"],
         "samples": [