+++ tests/meta/orders/test_meta_orders_lims.py
@@ -1,3 +1,4 @@
+from cg.constants import Pipeline
 from cg.meta.orders.lims import LimsHandler
 
 
@@ -99,7 +100,7 @@ def test_to_lims_balsamic(balsamic_order_to_submit):
     first_sample = samples[0]
     assert first_sample["name"] == "s1"
     assert set(sample["container"] for sample in samples) == set(["96 well plate"])
-    assert first_sample["udfs"]["data_analysis"] == "Balsamic"
+    assert first_sample["udfs"]["data_analysis"] == str(Pipeline.BALSAMIC)
     assert first_sample["udfs"]["application"] == "WGTPCFC030"
     assert first_sample["udfs"]["sex"] == "M"
     assert first_sample["udfs"]["family_name"] == "family1"